added Feedback tu user during processing of G-Code.
The M109 (Ste Temperature and wait) can take a very long time. With
wrong PID Settings or a target Temperatur below the environment
temperature it can last forever.

This change allowes the current temperature to be reportet back (like
Marlin does) while the M109 is executing. This way the user can see what
is going on.